feature setting 모듈로 setting 패키지에 있는 코드 이동
