+++ core/common/src/main/java/com/teampophory/pophory/common/navigation/NavigationProvider.kt
@@ -4,4 +4,5 @@ import android.content.Intent
 
 interface NavigationProvider {
     fun toOnboarding(): Intent
+    fun toLicense(): Intent
 }
