+++ core/designsystem/src/main/res/drawable/ic_chevron_down.xml
