+++ core/designsystem/src/main/res/drawable/ic_chevron_right.xml
