+++ feature/setting/build.gradle.kts
@@ -10,5 +10,16 @@ android {
 }
 
 dependencies {
+    // core module
+    implementation(projects.core.common)
+    implementation(projects.core.designsystem)
+
+    // domain module
     implementation(projects.domain.auth)
+
+    // feature module
+    implementation(projects.feature.auth)
+
+    implementation(libs.accompanist.webview)
+    implementation(libs.process.phoenix)
 }
