+++ feature/setting/src/main/java/com/teampophory/pophory/feature/setting/TeamScreen.kt
@@ -1,5 +1,6 @@
 package com.teampophory.pophory.feature.setting
 
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -17,19 +18,14 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.Dp
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
-import com.teampophory.pophory.R
 import com.teampophory.pophory.common.compose.DefaultPreview
 import com.teampophory.pophory.common.compose.bottomBorder
 import com.teampophory.pophory.designsystem.PophoryTheme
-import com.teampophory.pophory.feature.setting.component.SettingItem
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun TermScreen(
-    navController: NavController,
-    onNavigatePersonalTerms: () -> Unit = {},
-    onNavigateTerm: () -> Unit = {},
-    onNavigateOss: () -> Unit = {}
+fun TeamScreen(
+    navController: NavController
 ) {
     Scaffold(
         modifier = Modifier.fillMaxSize(),
@@ -37,7 +33,7 @@ fun TermScreen(
             CenterAlignedTopAppBar(
                 title = {
                     Text(
-                        text = "약관 및 정책",
+                        text = "포포리팀",
                         textAlign = TextAlign.Center,
                         modifier = Modifier.fillMaxWidth(),
                         style = PophoryTheme.typography.headline2
@@ -46,7 +42,7 @@ fun TermScreen(
                 navigationIcon = {
                     IconButton(onClick = { navController.popBackStack() }) {
                         Icon(
-                            painter = painterResource(id = R.drawable.ic_chevron_left),
+                            painter = painterResource(id = com.teampophory.pophory.designsystem.R.drawable.ic_chevron_left),
                             contentDescription = "Back To Home"
                         )
                     }
@@ -60,18 +56,19 @@ fun TermScreen(
                 .padding(it)
                 .fillMaxSize()
         ) {
-            SettingItem(title = "개인정보 처리방침", onClick = onNavigatePersonalTerms)
-            SettingItem(title = "이용 약관", onClick = onNavigateTerm)
-            SettingItem(title = "오픈소스 라이선스", onClick = onNavigateOss)
+            Image(
+                painter = painterResource(id = R.drawable.img_team),
+                contentDescription = "Team Image"
+            )
         }
     }
 }
 
 @DefaultPreview
 @Composable
-private fun TermScreenPreview() {
+private fun TeamScreenPreview() {
     PophoryTheme {
         val navController = rememberNavController()
-        TermScreen(navController)
+        TeamScreen(navController)
     }
 }
