+++ feature/setting/src/main/res/drawable-xhdpi/img_team.png
