+++ feature/setting/src/main/res/drawable-xxhdpi/img_team.png
