+++ feature/setting/src/main/res/drawable-xxxhdpi/img_team.png
